+++ backend/src/main/java/com/optimize/elykia/core/enumaration/AppNotificationType.java
@@ -1,0 +1,7 @@
+package com.optimize.elykia.core.enumaration;
+
+public enum AppNotificationType {
+    PAYMENT_DECLARATION,
+    CUSTOMER_ORDER,
+    TONTINE_CATCHUP
+}